+++ packages/@expo/cli/src/utils/codesigning.ts
@@ -11,7 +11,9 @@ import {
 import { ExpoConfig } from '@expo/config';
 import { getExpoHomeDirectory } from '@expo/config/build/getUserState';
 import JsonFile, { JSONObject } from '@expo/json-file';
+import { CombinedError } from '@urql/core';
 import { promises as fs } from 'fs';
+import { GraphQLError } from 'graphql';
 import { pki as PKI } from 'node-forge';
 import path from 'path';
 import { Dictionary, parseDictionary } from 'structured-headers';
@@ -381,7 +383,15 @@ async function fetchAndCacheNewDevelopmentCodeSigningInfoAsync(
   easProjectId: string
 ): Promise<CodeSigningInfo | null> {
   const actor = await ensureLoggedInAsync();
-  const app = await AppQuery.byIdAsync(easProjectId);
+  let app: AppByIdQuery['app']['byId'];
+  try {
+    app = await AppQuery.byIdAsync(easProjectId);
+  } catch (e) {
+    if (e instanceof GraphQLError || e instanceof CombinedError) {
+      return null;
+    }
+    throw e;
+  }
   if (!actorCanGetProjectDevelopmentCertificate(actor, app)) {
     return null;
   }